+++ router-tests/modules/custom-set-scopes/main.go
