+++ router-tests/modules/custom-set-scopes/module/module.go
